// jshandymanservicesoftallyinc-from/fromt/src/app/models/DataJshandyManServices.model.ts
export class DataJshandyManServices {

  idDataConfig: number;
  direction: string;
  mail: string;
  web: string;
  phoneNumber: string;
  taxRegNumber: string;
  coments1: string;
  coments2: string;
  coments3: string;
  coments4: string;
  portMail: string;
  company: string;
  deadLineAlert: number;
  active: Boolean;
  userCode: string;

  constructor(){
    this. active = true;
  }
}