// jshandymanservicesoftallyinc-from/fromt/src/app/models/loging-response.model.ts
import { DataJshandyManServices } from './DataJshandyManServices.model';

export class LogingResponse {
    authorization: string;
    token: string;
    status: any;
    username: string;
    userCode: string;
    taxes: number;
    userImagen: string;
    data: DataJshandyManServices;
    configurationRedy: boolean;
}